Optionally generate ID tokens
The ID tokens can be validated and used by the backends to authenticate
the users. These tokens contain the information that tlsproxy proxy
already verified with OIDC or SAML.
