+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/c2FmZQ/tlsproxy
 
-go 1.20
+go 1.21
 
 require (
 	github.com/beevik/etree v1.2.0
@@ -9,7 +9,6 @@ require (
 	github.com/golang-jwt/jwt/v5 v5.0.0
 	github.com/russellhaering/goxmldsig v1.4.0
 	golang.org/x/crypto v0.13.0
-	golang.org/x/exp v0.0.0-20230905200255-921286631fa9
 	golang.org/x/sys v0.12.0
 	golang.org/x/time v0.3.0
 	gopkg.in/yaml.v3 v3.0.1
